New project with osc controlled material parameters
Each Projection surface is an instance of ProjectionSurface - select the material you'd like it initialized to in the details pane, and scale and position as if it were any other static mesh.

Register it in the "ProjectionSurfaces" parameter of the OSCReceiver with a name (eg in the test projcect there already are Front, Back, Left, Right, Floor)

Then you can send osc messages of the form

`/UpdateMaterialParameter [surface name] [parameter] [value]`

eg `/UpdateMaterialParameter Front XCenter 0.5`

to update the parameter.

This will work for any parameter of the material without modification.

Limitations:

- This currently only listens for changes in material parameters. I'm going to add movements of each screen next which will be simple

- Each screen is initialized to a specific material and that can't be changed remotely. I think that could be added though, no problem.